+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud/compare/v1.0.0...HEAD)
 ------------------------------------------------------------------------
+### Changed
+- No longer analyses unversioned files
 
 [1.0.0](https://github.com/jshwi/pyaud/releases/tag/v1.0.0) - 2021-01-26
 ------------------------------------------------------------------------